fix: make use of private tokens possible again
Correct the token functionality of `gl-got` so that it sends the token
using the `PRIVATE-TOKEN` header.

During a refactor of `gl-got` the header was switched to use the
authentication header format used by GitHub.

However, GitLab uses a different header for personal access tokens.

Furthermore, OAuth2 uses a `Bearer` token in the `AUTHENTICATION`
header, instead of a plain `token`.

Because this package states that it supports Personal Access tokens
in the `README`, this change is not breaking. On the contrary, it
addresses a serious bug preventing it from behaving in the way
advertised.

Fixes #37